requirements.txt
aiohttp==3.7.4.post0
astroid==2.8.0
async-timeout==3.0.1
attrs==19.3.0
Automat==0.8.0
blinker==1.4
cachetools==4.2.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==21.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
discord.py==1.7.3
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
google-api-core==2.0.1
google-api-python-client==2.23.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.19.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
isort==5.9.3
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.2.0
multidict==5.1.0
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
platformdirs==2.4.0
protobuf==3.18.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.11.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.4
python-debian===0.1.36ubuntu1
python-dotenv==0.19.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.1
ssh-import-id==5.10
systemd-python==234
toml==0.10.2
Twisted==18.9.0
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
wadllib==1.3.3
wrapt==1.12.1
yarl==1.6.3
zipp==1.0.0
zope.interface==4.7.1
